requirements.txt
python-telegram-bot
requests
pymongo
dnspython